requirements.txt
aws-cdk-lib==2.63.2
constructs>=10.0.0,<11.0.0
attini-cdk-lib>=1.0.3
